v1.1.0: support for unit testing through stubbing
I've added support for unit testing through generating a 'stub' library
that resolves all the PROS symbols without actually linking to the PROS
library so that I can compile and link without errors.

Currently the stubs, as the name implies, do absolutely nothing, and
using the output of a stub, will result in undefined behaviour, though
that is fine, as the stub functions aren't meant to be actually called
in any of the unit tests, but rather to make the executable link and
run on non VexV5 Brain hardware.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>