v1.1.1: moved pros.zig & stub.zig into ./include
I've moved the pros.zig and stub.zig files into the include directory as
previusly, whenever you upgraded with the pros conductor command, it
would completely override the ./src dir and replace it; deleting any
user code along with it. This is due to the fact, that the src dir
was marked as belonging to this template and therefore wiped upon
upgrades. By moving the template's 'bindings' to the include dir, src
would no longer be wiped. It would also move the bindings out of the src
dir so the user wouldn't accidentally delete or modify them.

I've also added a module import to the userlib as otherwise they
wouldn't be able to import the pros.zig as it's no longer in the src dir

Signed-of-by: Ethan Ma <kalscium@protonmail.com>